Finished Melee Hit Detection & Added Target Filtering
Finished implementing the actual collision detection for melee strikes, as well as the ability to filter out actors that shouldn't be targetable.

Arguably that last point could be redundant since the unreal functions are already limiting hits to pawns, but it leaves space to modify friendly fire later on at least.