IK Retargeter
Set up an IK Retargeter to convert Mixamo animations to dungeon characters.
Might end up being kind of pointless for this particular project, but I know now how to do it so I'm happy with that.

Not set any IK rules up yet, just set up goals for the hands and feet so far.
Think that'll be enough for now but can revisit it later if the need arises.