Created Hit React Montage
Created a copy of the greatsword impact montage called hit react montage and populated it with all the stumbling animations.
I've made montage sections for each of them so that I can specify at runtime which of them I wish to play.
Not entirely sure how well it'll work with the state system at this point.