Death Half Implemented, ABP Needs Work
Imported and converted a whole bunch of death animations into the project.
I also set up a function that allows a character to die, as well as modifying how the behavior tree recognizes a dead character to respect the new changes.
Still having problems with the animation blueprint however; currently it doesn't recognize dead characters and so they stand right back up after finishing their death animation.

It should be noted that these changes only really apply to AI characters, as the only element that respects the dead really is the behaviour tree.
This means that the player can die and still move around but I'm not too concerned about this since we might throw out my character code for Ryan's when we merge our work soon.