Line & Sphere Draw, Component Refs Crashing Editor
Modified the melee component to be a bit more flexible, it can now draw a line between two points or a sphere around a single point.
I'm currently attempting to join sphere components in blueprint with my C++ code. Using component references seems to be one possible solution, but it also seems to be crashing the editor so I'm saving here before I continue.